:ok_hand: Split utils.jsx into two files, decouple functions
