+++ web/src/hitsByIncomingLink.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { List, Datagrid, FunctionField, NumberField } from "react-admin";
-import { toAbsoluteIncomingLink } from "./utils";
+import { toAbsoluteIncomingLink } from "./linkFormatting";
+import { toLinkOpeningNewTab } from "./renderUrl";
 
 export const HitsByIncomingLinkList = (props) => (
   <List {...props}>
@@ -8,7 +9,9 @@ export const HitsByIncomingLinkList = (props) => (
       <FunctionField
         source="incoming_link"
         label="Incoming Link"
-        render={(record) => toAbsoluteIncomingLink(record.incoming_link)}
+        render={(record) =>
+          toLinkOpeningNewTab(toAbsoluteIncomingLink(record.incoming_link))
+        }
       />
       <NumberField source="hit_count" label="Click count" />
     </Datagrid>
